Allow Terminal Arguments
This allows us to create the terminal with arguments.  The immediate use
case is to open an ipython shell with the `--no-autoindent` flag which
means that we can use the shortcuts to send lines to the attached
terminal without getting double indents and thus failing to execute an
indenent sensitive language like python.