Bug: Python Premature Ejeculution
When running python visual selection with a newline in the middle, the
first bit gets executed without continuing to the second bit.

This solution just strips out empty lines in the input so that the shell
doesn't get the execute signal until the end.

We also add an extra newline at the end of the complete command to
ensure python code runs since ipython shell will let you enter a solo
newline and continue if you are dropping out of a loop.