Fix bash kemap to 'clear' the terminal
This seems to send the terminal buffer into normal mode when in neovim
(normal vim it works as expected).

I relise we have no consistent philosoply with what keymaps to keep vim
and what bash ones to use but this is a common one for me. Much more
common than the <C-w> in bash to remove the word before the cursor.